% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brewin_es}
\alias{brewin_es}
\title{Brewin & Field (2024) with effect sizes}
\format{
A tibble with 80 rows and 19 variables.
}
\source{
\url{https://osf.io/597hr/}
}
\usage{
brewin_es
}
\description{
Data from a meta-analysis looking at predictors of disorganization in trauma memories. This is a version of the data in \link{brewin_2024}
where the effect sizes have already been computed and some non-essential variables are omitted.
}
\details{
Background: Taylor et al. (Clinical Psychological Science, 2022) reported that in healthy participants memories of
traumatic and comparison films did not differ in coherence. The lack of a group diagnosed with post-traumatic stress disorder (PTSD),
as well as limitations of the trauma film paradigm, mean that their design is unable to directly test predictions made by clinical theories of PTSD. Brewin and Field argued that
there is convincing evidence for trauma memories in PTSD being incoherent or disorganized and conducted a meta-analysis to estimate the effect size for PTSD status and memory incoherence/disorganization.
These are the raw data from this meta-analysis. Key to the analysis is the fact that there are two distinct approaches to measuring disorganization/incoherence with
acceptable face validity. One, based on therapy with PTSD patients, is to elicit a very detailed trauma narrative that includes the worst moments and have judges rate individual utterance units for markers of disorganization such as repetition and non-consecutive chunks
(The FOA method). The other approach, employed by Taylor et al. (2022) among others, has judges or participants rate the entire memory or narrative. The tibble contains the following variables
\itemize{
\item \strong{author}: the author of the paper
\item \strong{study}: numeric study identifier
\item \strong{high_disorganization}: is high disorganization represented by a low score or a high score
\item \strong{clinical_group}: Did the clinical group have diagnosis of PTSD or ASD
\item \strong{age}:	Adult or youth sample?
\item \strong{clinical_mean}:	mean disorganization in the clinical group
\item \strong{clinical_sd}:	standard deviation of disorganization in the clinical group
\item \strong{clinical_n}:	sample size disorganization in the clinical group
\item \strong{control_mean}: mean disorganization in the control group
\item \strong{control_sd}:	standard deviation of disorganization in the control group
\item \strong{control_n}:	sample size disorganization in the control group
\item \strong{es_id}:	numeric identifier for unique effect sizes
\item \strong{foa_gp}: a categorical variable combining information about the \code{foa_type} and \code{foa} group.
\item \strong{global_vs_detail}: the first of three dummy variables. This one compares global FOA methodology to detailed FOA methodology
\item \strong{dis_vs_detail}:	the second of three dummy variables. This one compares non-FOA methodology that measures disorganization to detailed FOA methodology
\item \strong{org_vs_detail}: the last of three dummy variables. This one compares non-FOA methodology that measures organization to detailed FOA methodology
\item \strong{g}: Hedge's g
\item \strong{v_g}: variance for Hedge's g
\item \strong{n_total}: total sample size
}
}
\references{
\itemize{
\item Brewin, C. & Field, A. P. (2024). Meta-analysis shows trauma memories in PTSD lack coherence: A response to Taylor et al. (2022). Clinical Psychological Science. \url{https://osf.io/597hr/}
\item Taylor, A., Zajac, R., Takarangi, M. K. T., & Garry, M. (2022). Evidence from the trauma-film paradigm that traumatic and nontraumatic memories are statistically equivalent on coherence. \emph{Clinical Psychological Science}, 10(3), 417–429. \url{https://doi.org/10.1177/21677026211053312}
}
}
\keyword{datasets}
