% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{forest_add_het}
\alias{forest_add_het}
\title{Add heterogeneity statistics to a forest plot}
\usage{
forest_add_het(text = "RE model:", model, dp = 2)
}
\arguments{
\item{text}{a text string used as a prefix. Default is "RE model:"}

\item{model}{A meta-analyses object  created using \code{metafor::rma()}}

\item{dp}{number of decimal places to print in the output}
}
\value{
Text to be added to a forest plot.
}
\description{
\code{forest_add_het_stats()} is a helper function to add heterogeneity statistics to a forest plot. This function is adapted from \url{https://www.metafor-project.org/doku.php/plots:forest_plot_with_subgroups}
}
