% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{forest_subgroups}
\alias{forest_subgroups}
\title{Forest plot for levels of a factor}
\usage{
forest_subgroups(
  tibble,
  es,
  var_es,
  groups,
  slab,
  xlim,
  at,
  dp = 2,
  author_lab = "Study",
  gp_labs = NULL,
  cex = 0.75
)
}
\arguments{
\item{tibble}{A tibble containing effect sizes and their variances.}

\item{es}{The name of the variable containing the effect sizes}

\item{var_es}{The name of the variable containing the variance of each effect size}

\item{groups}{The name of the factor variable for which you want to split the plot}

\item{slab}{The name of the variable containing study labels (this variable is used to set the \code{slab} argument within \code{forest()} )}

\item{xlim}{Two numeric values specifying the lower and upper limit of the x-axis (e.g. \code{xlim = c(-2.5, 2.5)})}

\item{at}{Numeric values to be used to set the \code{at} argument within \code{forest()} (e.g., \code{at = seq(-2.5, 2.5, 0.5)})}

\item{dp}{Default is 2 decimal places.}

\item{author_lab}{The text label to use at the top of the right hand side of the plot (by default, \code{author_lab = "Study"})}

\item{gp_labs}{By default the factor labels are used to label the subgroups. This argument overrides this behavior if you supply a vector of labels. If you override the default behaviour bear in mind that \code{forest()} prints effect sizes from the bottom up so you need to order your labels accordingly: your first label should be for the group of studies plotted at the bottom of the plot and the last label should be for the subgroup plotted at the top of the plot.}

\item{cex}{A numeric value to be used to set the \code{cex} argument within \code{forest()} (by default \code{cex = 0.75})}
}
\value{
A forest plot split by subgroups.
}
\description{
\code{forest_subgroups()} plots a forest plot where effect sizes are split by subcategories of a factor. Essentially it attempts to create
\url{https://www.metafor-project.org/doku.php/plots:forest_plot_with_subgroups}
programatically.
}
