% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{get_mod_mas}
\alias{get_mod_mas}
\title{Run individual meta-analyses with a moderator variable for each level of a categorical predictor.}
\usage{
get_mod_mas(
  tibble,
  predictor,
  moderator,
  es_id,
  study_id,
  digits = 2,
  p_digits = 3,
  summary = TRUE
)
}
\arguments{
\item{tibble}{A tibble containing the effect sizes, their variances,a study ID and an effect size ID}

\item{predictor}{Name of the categorical variable for which you want individual meta-analyses}

\item{moderator}{Name of the moderator variable to be included within each sub-analysis}

\item{es_id}{Name of the variable that identifies unique studies}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}

\item{summary}{if TRUE output a summary table, otherwise store the raw data and models}
}
\value{
A tibble containing the results of the meta-analysis for each category in a separate row.
Columns show
\itemize{
\item \emph{k}: number of effect sizes
\item \eqn{\sigma_b}: between study variance
\item \eqn{\sigma_w}: within-study variance
\item \emph{q}: Q statistic for heterogeneity,
\item \eqn{q_{df}}: degrees of freedom for the \emph{Q} statistic,
\item \eqn{q_p}: \emph{p}-value for the Q statistic
\item estimate: The estimate of the effect of the moderator
\item ci: the 95\% confidence interval for the estimate of the effect of the moderator
\item statistic: the test statistic for the estimate of the effect of the moderator
\item p.value: the test statistic for the estimate of the effect of the moderator
}
}
\description{
\code{get_mod_mas()} does the same thing as \code{\link[=get_mas]{get_mas()}} but allows the user to specify a moderator/predictor within the individual meta-analyses.
So it fits individual meta-analyses models with a moderator specified using the \code{rma.mv()} function from the \code{metafor} package
within each level of a categorical variable. Optionally, the results can be collated into a tabulated form for printing in a quarto document.
}
