% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metahelpr-package.R
\docType{package}
\name{metahelpr}
\alias{metahelpr-package}
\alias{metahelpr}
\title{metahelpr: Tutorials and helper functions for meta-analysis using metafor}
\description{
The goal of \code{metahelpr} is to offer tutorials and helper functions for conducting meta-analysis using the \code{metafor} package (Viechtbauer, 2010).
}
\section{Interactive tutorials}{


\strong{Getting started}:

I recommend working through \href{https://youtube.com/playlist?list=PLEzw67WWDg83weG3idsgy4wuOIJAashA2&si=PiI-sDvqc1DkaWOq}{this playlist of tutorials} on how to install, set up and work within R, RStudio and Quarto before starting the interactive tutorials.

\strong{Running a tutorial}:

To run each tutorial execute

\code{learnr::run_tutorial("name_of_tutorial", package = "metahelpr")}

Replacing \code{name_of_tutorial} with the name in bold below. For example, to load the tutorial \code{meta_d} execute:

\code{learnr::run_tutorial("meta_d", package = "metahelpr")}
\itemize{
\item \strong{meta_d}: A tutorial on using the \href{https://wviechtb.github.io/metafor/}{\code{metafor} package} to conduct a meta-analysis using Cohen's \eqn{\hat{d}} as the effect size measure.
\item \strong{meta_r}: A tutorial on using the \href{https://wviechtb.github.io/metafor/}{\code{metafor} package} to conduct a meta-analysis using Pearson's $r$ as the effect size measure.
}
}

\section{Datasets}{

\itemize{
\item \link{brewin_2024}: Raw data from the meta-analysis by Brewin & Field (2024).
\item \link{brewin_es}: Reduced dataset from the meta-analysis by Brewin & Field (2024) that includes effect sizes.
\item \link{pearce_2016}: A selection of variables from the dataset from the meta-analysis by Pearce & Field (2016).
}
}

\section{Helper functions}{


The package contains some (probably badly written and easily breakable) helper functions that I often use when I conduct meta-analysis.
\itemize{
\item \link{d_from_r}: estimate Cohen's \emph{d} based on a biserial correlation coefficient, or when a biserial correlation isn't available the conversion uses the method described by  Mathur & VanderWeele (2019).
\item \link{d_to_g}: converts Cohen's \emph{d} to the unbiased Hedges' \emph{g}.
\item \link{forest_add_het_stats}: Add heterogeneity statistics to a forest plot
\item \link{forest_subgroups}: plots a forest plot where effect sizes are split by subcategories of a factor.
\item \link{get_mas}: Fit individual meta-analyses for each level of a categorical predictor and (optionally) collate the results into a tabulated form for printing in a quarto document.
\item \link{get_mod_mas}: does the same thing as \link{get_mas} but allows the user to specify a moderator/predictor within the individual meta-analyses. So, it fits individual meta-analyses models with a moderator specified using the \code{rma.mv()} function from the \code{metafor} package within each level of a categorical variable. Optionally, the results can be collated into a tabulated form for printing in a quarto document.
\item \link{get_pbm}: fits and collates publication bias models across categories of a predictor variable. It is assumed that you will supply two vectors of values one representing moderate publication bias and the other representing severe.
\item \link{plot_bubble}: create a bubble plot (using \code{ggplot2}) based on a \code{regtest()} object.
\item \link{pooled_var}: computes a full sample variance estimate based on means and variances from two subgroups.
\item \link{regtest_tbl}: puts the results of \code{metafor::regtest()} into a tibble for reporting.
\item \link{report_het}: collates information from heterogeneity tests and outputs text that summarizes the results in a format that will render nicely in quarto.
\item \link{report_mod}: outputs text that reports (and renders nicely in quarto/Rmarkdown) the omnibus statistical tests from a moderation model.
\item \link{report_par_tbl}: outputs a tibble of the table of coefficients of a meta-analyses object (\code{rma} or \code{rma.mv}) created using the \code{metafor} package. This function will mostly be useful for models containing predictors of effect sizes (so-called meta-regression).
\item \link{report_pars}: outputs text that reports (and renders nicely in quarto/Rmarkdown) the individual effects from a meta-analysis model.
\item \link{var_d_from_r}: Estimate the sampling variance of Cohen's \emph{d} based on a correlation coefficient
}
}

\section{References}{

\itemize{
\item Brewin, C. R., & Field, A. P. (2024). Meta-analysis shows trauma memories in PTSD lack coherence: A response to Taylor et al. (2022). \emph{Clinical Psychological Science}.
\item Mathur, M. B., & VanderWeele, T. J. (2020). A simple, interpretable conversion from Pearson's correlation to Cohen's \emph{d} for continuous exposures. \emph{Epidemiology}, 31(2), e16–e17. \href{https://doi.org/10.1097/EDE.0000000000001111}{doi.org/10.1097/EDE.0000000000001111}
\item Pearce, L. J., & Field, A. P. (2016). The impact of 'scary' TV and film on children’s internalizing emotions: A meta-analysis. \emph{Human Communication Research}, 42, 98–121. \url{https://doi.org/doi:10.1111/hcre.12069}
\item Viechtbauer, W. (2010). Conducting Meta-Analyses in R with the metafor Package. \emph{Journal of Statistical Software}, 36, 1–48.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/profandyfield/metahelpr}
  \item Report bugs at \url{https://github.com/profandyfield/metahelpr/issues}
}

}
\author{
\strong{Maintainer}: Andy Field \email{andyf@sussex.ac.uk}

}
\keyword{internal}
