% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pearce_2016}
\alias{pearce_2016}
\title{Pearce & Field (2016)}
\format{
A tibble with 129 rows and 19 variables.
}
\source{
\url{https://doi.org/doi:10.1111/hcre.12069}
}
\usage{
pearce_2016
}
\description{
Background: Despite a general perception that violent or scary television creates anxiety in children, the research literature was small and disparate.
In 2016 we conducted a meta-analysis that quantified the impact of scary television and film on children's internalizing emotions (fear, anxiety, sadness, and sleep problems).
\code{pearce_2016} contains the key variables from the data used in this study. The data contains the following variables
}
\details{
\itemize{
\item \strong{author}: the author of the paper
\item \strong{study_id}: numeric study identifier
\item \strong{es_id}:	numeric identifier for unique effect sizes
\item \strong{year}: year of publication
\item \strong{methodology}: Did the clinical group have diagnosis of PTSD or ASD
\item \strong{responder}: was the responder the child, the parent or both
\item \strong{age_range_cat}: age range of participants (text)
\item \strong{age_range_num}:	age range of participants (numeric)
\item \strong{mean_age_est}:	estimated mean age of participants
\item \strong{all_under_10}: was all of the sample under 10 years old?
\item \strong{measure_type}: Type of measure used (e.g. self-report)
\item \strong{outcome}: Outcome measured (fear, PTSD, general, Depression/sadness, physiology, sleep disorder)
\item \strong{news_fiction}: was the media consumed news or fiction?
\item \strong{r}: effect size \emph{r}
\item \strong{var_r}:	variance of the effect size
\item \strong{n}:	total sample size onw hich the effect size is based
\item \strong{tv}: was the media consumed television or mixed media?
\item \strong{content}: classification of the content consumed (e.g. terrorism)
\item \strong{violence}: rating of how violent the content was from 0 (not at all violent) to 3 (most violent)
}
}
\references{
\itemize{
\item Pearce, L. J., & Field, A. P. (2016). The impact of 'scary' TV and film on children’s internalizing emotions: A meta-analysis. \emph{Human Communication Research}, 42, 98–121. \url{https://doi.org/doi:10.1111/hcre.12069}
}
}
\keyword{datasets}
