% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{plot_bubble}
\alias{plot_bubble}
\title{Bubble plots from a \code{regtest()} object}
\usage{
plot_bubble(
  model,
  xlim = NULL,
  ylim = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  xlab = "Predictor",
  ylab = "Effect size",
  title = NULL,
  pi = TRUE
)
}
\arguments{
\item{model}{A meta-analyses object (\code{rma} or \code{rma.mv}) created using the \code{metafor} package}

\item{xlim}{Limits for the \emph{x}-axis}

\item{ylim}{Limits for the \emph{y}-axis}

\item{xbreaks}{Breaks for the \emph{x}-axis}

\item{ybreaks}{Breaks for the \emph{y}-axis}

\item{xlab}{Text label for the \emph{x}-axis}

\item{ylab}{Text label for the \emph{y}-axis}

\item{title}{Text title for the plot}

\item{pi}{If \code{TRUE} uses a 95\% prediction interval, if \code{FALSE} uses a 95\% confidence interval}
}
\value{
A tibble containing (for each predictor in the model) the name of the predictor (obtained from \code{pred_labels}), the effect size estimate, 95\% confidence interval, test statistic and \emph{p}-value.
}
\description{
\code{plot_bubble()} is a helper function to create a bubble plot (using \code{ggplot2}) based on a \code{regtest()} object.
}
