% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{pooled_var}
\alias{pooled_var}
\title{Compute the full sample variance/SD from the means, SDs and sample sizes of two subgroups}
\usage{
pooled_var(nx, ny, sdx, sdy, ux, uy, sd = F)
}
\arguments{
\item{nx}{numeric value of the sample size for group 1}

\item{ny}{numeric value of the sample size for group 2}

\item{sdx}{numeric value of the standard deviation for group 1}

\item{sdy}{numeric value of the standard deviation for group 2}

\item{ux}{numeric value of the mean for group 1}

\item{uy}{numeric value of the mean for group 2}

\item{sd}{If \code{FALSE} returns the variance, if \code{TRUE} returns the standard deviation}
}
\value{
The combined standard deviation or variance of the two groups.
}
\description{
\code{pooled_var()} is a helper function that computes a full sample variance estimate based on means and variances from two subgroups. This is done using the following equation
\deqn{s^2_\text{combined} = \frac{1}{n_1 + n_2 -1}\left((n_1-1)s_1^2 + (n_2-1)s_2^2 + \frac{n_1n_2}{n_1+n_2}(\overline{x}_1-\overline{x}_2)^2\right)}
This equation is essentially the same as \emph{Result 1} in O'Neill (2014).
}
\references{
O’Neill, B. (2014). Some useful moment results in sampling problems. \emph{The American Statistician}, 68(4), 282–296. \url{https://doi.org/10.1080/00031305.2014.966589}
}
