% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{regtest_tbl}
\alias{regtest_tbl}
\title{Produce tabulated output for regression tests}
\usage{
regtest_tbl(model, digits = 3, p_digits = 3)
}
\arguments{
\item{model}{A model created with \code{rma()} from the \code{metafor} package.}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}
}
\value{
A tibble containing the results of the regression test.
Columns show
\itemize{
\item estimate: The limit estimate for the model
\item ci: the 95\% confidence interval for the limit estimate for the model
\item statistic: the test statistic for the limit estimate for the model
\item p.value: the test statistic for the limit estimate for the model
}
}
\description{
\code{regtest_tbl()} puts the results of \code{metafor::regtest()} into a tibble for reporting.
}
