% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{report_het}
\alias{report_het}
\title{Report residual heterogeneity statistics}
\usage{
report_het(metaobject, digits = 2, p_digits = 3)
}
\arguments{
\item{metaobject}{A meta-analyses object (\code{rma} or \code{rma.mv}) created using the \code{metafor} package}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}
}
\value{
A text string including LaTeX formatting for mathematical symbols that reports the \emph{Q} statistic, its degrees of freedom and it's associated \emph{p}-value.
}
\description{
\code{report_het()} is a helper function that collates information from heterogeneity tests and outputs text that summarizes the results in
a format that will render in quarto. Basically it helps you to report results.
}
