% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{report_mod}
\alias{report_mod}
\title{Report residual heterogeneity statistics}
\usage{
report_mod(metaobject, digits = 2, p_digits = 3)
}
\arguments{
\item{metaobject}{A meta-analyses object (\code{rma} or \code{rma.mv}) created using the \code{metafor} package}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}
}
\value{
A text string including LaTeX formatting for mathematical symbols that reports the \emph{Q} statistic (or \emph{F} for robust models), the associated degrees of freedom and \emph{p}-value.
}
\description{
\code{report_mod()} is a helper function that outputs text that reports the omnibus statistical tests from a moderation model.
Basically it helps you to report results in quarto.
}
