% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{report_pars}
\alias{report_pars}
\title{Report parameters from a meta-analysis object}
\usage{
report_pars(metaobject, row = 1, digits = 2, p_digits = 3, mod = F)
}
\arguments{
\item{metaobject}{A meta-analyses object (\code{rma} or \code{rma.mv}) created using the \code{metafor} package}

\item{row}{The row of output you wish to report (if you have more than one effect)}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}

\item{mod}{Are you reporting the effect of a moderator variable? This affects the symbols used for the parameter estimates. For moderators the
symbol \eqn{\hat{\beta}} is used, for overall effects \eqn{\hat{\theta}} is used.}
}
\value{
A text string including LaTeX formatting for mathematical symbols that reports the estimate (\eqn{\hat{\beta}} for moderators, \eqn{\hat{\theta}} for overall effects),
95\% confidence interval, associated test-statistic and \emph{p}-value.
}
\description{
\code{report_pars()} is a helper function that outputs text that reports the individual effects from a meta-analysis model.
Basically it helps you to report results in quarto.
}
