% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{var_d_from_r}
\alias{var_d_from_r}
\title{Estimate the sampling variance of Cohen's \emph{d} based on a correlation coefficient}
\usage{
var_d_from_r(n, r, d, var = T)
}
\arguments{
\item{n}{the total sample size on which \emph{d} was based}

\item{r}{The correlation coefficient on which \emph{d} was based}

\item{d}{The estimate of \emph{d}. If \code{d} is missing (the default) the variance us calculated assuming \emph{r} is a biserial correlation. If \emph{d} is set to a value then the function
treats the measures as continuous based on Mathur & VanderWeele (2019).}

\item{var}{When \code{TRUE} the sampling variance is returned, when \code{FALSE} the standard error is returned.}
}
\value{
Sampling variance or standard error of Cohen's \emph{d}
}
\description{
When the raw information is not available to compute \emph{d}, #' it may be necessary to find it indirectly.
These conversions are approximations only. \code{\link[=d_from_r]{d_from_r()}} estimates Cohen's \emph{d} from a correlation coefficient and \code{var_d_from_r()} estimates the
associated sampling variance. When a value of \code{d} is provided the function uses Mathur & VanderWeele (2019) to calculate the corresponding variance,
otherwise (if d is missing) the equation based on the biserial correlation is used.
}
\references{
Mathur, M. B., & VanderWeele, T. J. (2020). A simple, interpretable conversion from Pearson's correlation to Cohen's \emph{d} for continuous exposures. \emph{Epidemiology}, 31(2), e16–e17. \url{https://doi.org/10.1097/EDE.0000000000001111}
}
