% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R
\name{get_pbm}
\alias{get_pbm}
\title{Publication bias weight models}
\usage{
get_pbm(
  tibble,
  predictor,
  es,
  var_es,
  digits = 2,
  p_digits = 3,
  a,
  moderate,
  severe
)
}
\arguments{
\item{tibble}{A tibble containing the raw data}

\item{predictor}{The name of the categorical predictor for which you want individual weight models (optional)}

\item{es}{Name of the variable containing the effect sizes}

\item{var_es}{Name of the variable containing the variance estimate of the effect sizes}

\item{digits}{number of decimal places to print in the output}

\item{p_digits}{number of decimal places for \emph{p}-values}

\item{a}{numeric vector of one or more values that will be used to set the \code{steps} argument in \code{metafor::selmodel()}}

\item{moderate}{A vector of weights for moderate publication bias (eg. \code{c(1, .7, .6, .5)}) that will be used to set the \code{delta} argument in \code{metafor::selmodel()}}

\item{severe}{A vector of weights for severe publication bias (eg. \code{c(1, .5, .4, .1)}) that will be used to set the \code{delta} argument in \code{metafor::selmodel()}}
}
\value{
A tibble containing (for each category of the predictor) a row reporting the moderate publication bias results and another reporting the severe. The table contains the
original effect size estimate (\eqn{\hat{\theta}}) the estimate adjusted for bias ((\eqn{\hat{\theta}_{\text{Adj}}})), the 95\% confidence interval for the adjusted value and its associated \emph{p}-value.
This tibble can be printed using \code{kable()} in quarto.
}
\description{
\code{get_pbm()} is a helper function to fit and collate publication bias models. It is assumed that you will supply
two vectors of values one representing moderate publication bias and the other representing severe. If a categorical variable is
supplied to \code{predictor} then the function will fit and collate publication bias models across categories of that variable.
For more information see \code{\link[metafor:selmodel]{metafor::selmodel()}}
}
